added eval tests

+++ src/aqmodels/_solution.pyi
@@ -53,7 +53,7 @@ class Solution:
         binary_cols: list[list[int]] | None = ...,
         spin_cols: list[list[int]] | None = ...,
         int_cols: list[list[int]] | None = ...,
-        real_cols: list[list[int]] | None = ...,
+        real_cols: list[list[float]] | None = ...,
         raw_energies: list[float | None] | None = ...,
         timing: Timing | None = ...,
     ) -> Solution: ...
